+++ R/data.R
@@ -1,6 +1,7 @@
 #' Simulated lung cancer patients data by Maringe et al. (2020)
 #'
-#' The dataset is from supplmentary files of [Maringe et al. (2020)](https://doi.org/10.1093/ije/dyaa057).
+#' The dataset is from the supplementary files of Maringe et al. (2020),
+#' \doi{10.1093/ije/dyaa057}.
 #' The dataset is a set of 200 simulated lung cancer patients. 
 #' These patients are followed up for a year following their cancer diagnosis:
 #' 106 of them received surgery within six months of their diagnosis and 
@@ -22,7 +23,7 @@
 #'   \item{charlson}{Charlson's comorbidity index}
 #'   \item{emergency}{route to diagnosis}
 #' }
-#' @source <https://doi.org/10.1093/ije/dyaa057>
+#' @source \doi{10.1093/ije/dyaa057}
 #' @examples
 #' data(lungcancer)
 "lungcancer"
@@ -30,7 +31,8 @@
 
 #' 13 types of patient records in Maringe et al. (2020)
 #'
-#' The dataset is from Figure 2 in [Maringe et al. (2020)](https://doi.org/10.1093/ije/dyaa057).
+#' The dataset is from Figure 2 in Maringe et al. (2020),
+#' \doi{10.1093/ije/dyaa057}.
 #' The dataset illustrates all possible censoring mechanisms with 13 types of 
 #' patients records that could be seen in the cancer registry data, when 
 #' allowing at most one treatment (surgery in this example) for each patient
@@ -46,7 +48,7 @@
 #'   \item{death}{observed event at the latest follow-up, 1: dead, 0: alive}
 #'   \item{followup}{observed follow-up time (time to death or time to latest followup with 1 year (365 days) at maximum)}
 #' }
-#' @source <https://doi.org/10.1093/ije/dyaa057>
+#' @source \doi{10.1093/ije/dyaa057}
 #' @examples
 #' data(patients13)
 "patients13"