+++ tests/testthat/test-read_trial_data.R
@@ -15,3 +15,9 @@ test_that("read_trial_data reads csv input into a tibble", {
   expect_equal(names(result), c("id", "follow_up", "event", "treatment"))
   expect_equal(nrow(result), 2)
 })
+
+
+
+
+.libPaths("/Users/sanghopark/Library/Caches/org.R-project.R/R/renv/sandbox/macos/R-4.4/aarch64-apple-darwin20/f7156815"      )
+